markaz/requirement.txt
amqp==5.1.1
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==4.1.0
Brotli==1.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.4
cssselect2==0.7.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-apscheduler==0.6.2
django-captcha==1.0.0
django-import-export==3.2.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.42.1
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.2
lxml==4.9.3
MarkupPy==1.14
mysql-connector-python==8.1.0
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
Pillow==10.0.1
prompt-toolkit==3.0.39
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.7.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pyOpenSSL==23.2.0
pyotp==2.9.0
pypdf==3.16.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
reportlab==3.6.13
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
twisted==23.8.0
txaio==23.1.1
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.1
zopfli==0.2.3
